requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests
uvloop
scikit-learn
numpy
pandas
joblib
tensorflow==2.12.0
websockets
aiohttp
